+++ src/server/js/functions.js
@@ -6,19 +6,108 @@ import _ from 'lodash';
 const assert = require( 'assert' ).strict;
 
 import { Logger } from 'meteor/pwix:logger';
+import { Mongo } from 'meteor/mongo';
 import { Roles as alRoles } from 'meteor/alanning:roles';
 
 const logger = Logger.get();
 
 Roles.s = {
+    // get all roles for the user
+    // returns {Object} an object with following keys:
+    //   - global: the global roles as an object with following keys:
+    //     > all: an array of all roles
+    //     > direct: an array of direct roles
+    //   - scoped: the scoped roles as an object keyed by the scope identifier, with folloging keys
+    //     > all: an array of all roles for this scope
+    //     > direct: an array of direct roles for this scope
+    async allRolesForUser( target, requester=null ){
+        if( !target || ( !_.isString( target ) && !_.isObject( target ))){
+            logger.error( 'allRolesForUser() expect target be a user identifier or a user document, got', target, 'throwing...' );
+            throw new Error( 'Bad data type' );
+        }
+        if( !requester || ( !_.isString( requester ) && !_.isObject( requester ))){
+            logger.error( 'allRolesForUser() expect requester be a user identifier or a user document, got', requester, 'throwing...' );
+            throw new Error( 'Bad data type' );
+        }
+        try {
+            const allowed = await Roles.isAllowed( 'pwix.roles.fn.getRolesForUser', requester, target );
+            if( allowed ){
+                const collectionName = Roles.configure().assignmentsCollection;
+                const collection = Mongo.getCollection( collectionName );
+                if( !collection || !( collection instanceof Mongo.Collection )){
+                    logger.error( 'allRolesForUser() expect collection \''+collectionName+'\' be an instance of Mongo.Collection, got', collection, 'throwing...' );
+                    throw new Error( 'Bad data type' );
+                }
+                let targetId = target;
+                if( _.isObject( target )){
+                    targetId = target._id;
+                }
+                const fetched = await collection.find({ 'user._id': targetId }).fetchAsync();
+                //logger.debug( 'fetched', fetched );
+                // get something like:
+                //    fetched [
+                //    {
+                //        _id: 'Wm64TCbu2AwjRtuDc',
+                //        user: { _id: 'AtJ9dNdzumE5PkPA5' },
+                //        role: { _id: 'APP_ADMINISTRATOR' },
+                //        scope: null,
+                //        inheritedRoles: [
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object], [Object], [Object],
+                //        [Object], [Object]
+                //        ]
+                //    },
+                //    {
+                //        _id: 'ED8XMfiqMuoPgdD5j',
+                //        user: { _id: 'AtJ9dNdzumE5PkPA5' },
+                //        role: { _id: 'SCOPED_IDENTITIES_MANAGER' },
+                //        scope: 'bDZcDsWtuqyJJAQcb',
+                //        inheritedRoles: [
+                //        [Object], [Object],
+                //        [Object], [Object],
+                //        [Object], [Object],
+                //        [Object], [Object],
+                //        [Object]
+                //        ]
+                //    }
+                //  ]
+                let roles = { global: { direct: [], all: [] }, scoped: {}};
+                for( const it of fetched ){
+                    if( it.scope ){
+                        roles.scoped[it.scope] = roles.scoped[it.scope] || { all: [], direct: [] };
+                        Roles._doSetup( it, roles.scoped[it.scope] );
+                    } else {
+                        Roles._doSetup( it, roles.global );
+                    }
+                }
+                return roles;
+            }
+            return null;
+        }
+        catch( e ){
+            logger.error( 'allRolesForUser()', e );
+            return null;
+        }
+    },
+
     // get roles for the user
     // https://meteor-community-packages.github.io/meteor-roles/classes/Roles.html#method_getRolesForUserAsync
     // returns Promise null if an error occurred
     async getRolesForUser( user, options, userId=null ){
+        logger.warn( 'getRolesForUser() is obsoleted started with v1.9. Please use allRolesForUser()' );
         try {
             const allowed = await Roles.isAllowed( 'pwix.roles.fn.getRolesForUser', userId, user );
             if( allowed ){
-                return await alRoles.getRolesForUserAsync( user, options );
+                const roles = await alRoles.getRolesForUserAsync( user, options );
+                logger.debug( 'getRolesForUser()', options, roles );
             }
             //logger.log( 'getRolesForUser', user, 'not allowed' );
             return null;